+++ optaplanner-benchmark/src/main/java/org/optaplanner/benchmark/impl/result/SolverBenchmarkResult.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -337,11 +337,11 @@ public class SolverBenchmarkResult {
         if (!firstNonFailure) {
             int successCount = getSuccessCount();
             averageScore = totalScore.divide(successCount);
-            averageWorstScoreDifferencePercentage = totalWorstScoreDifferencePercentage.divide((double) successCount);
-            averageScoreCalculationSpeed = totalScoreCalculationSpeed / (long) successCount;
-            averageTimeMillisSpent = totalTimeMillisSpent / (long) successCount;
+            averageWorstScoreDifferencePercentage = totalWorstScoreDifferencePercentage.divide(successCount);
+            averageScoreCalculationSpeed = totalScoreCalculationSpeed / successCount;
+            averageTimeMillisSpent = totalTimeMillisSpent / successCount;
             averageWorstScoreCalculationSpeedDifferencePercentage = totalWorstScoreCalculationSpeedDifferencePercentage
-                    / ((double) successCount);
+                    / successCount;
         }
     }
