+++ optaplanner-core/src/main/java/org/optaplanner/core/api/score/buildin/simple/SimpleScore.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -169,7 +169,7 @@ public final class SimpleScore extends AbstractScore<SimpleScore> {
 
     @Override
     public String toShortString() {
-        return buildShortString((n) -> ((Integer) n).intValue() != 0, "");
+        return buildShortString((n) -> n.intValue() != 0, "");
     }
 
     @Override
