+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/heuristic/selector/move/decorator/AbstractCachingMoveSelector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -67,7 +67,7 @@ public abstract class AbstractCachingMoveSelector<Solution_> extends AbstractMov
     @Override
     public void constructCache(SolverScope<Solution_> solverScope) {
         long childSize = childMoveSelector.getSize();
-        if (childSize > (long) Integer.MAX_VALUE) {
+        if (childSize > Integer.MAX_VALUE) {
             throw new IllegalStateException("The selector (" + this
                     + ") has a childMoveSelector (" + childMoveSelector
                     + ") with childSize (" + childSize
